requirements.txt
eel==0.16.0
Pillow==10.1.0